pyproject.toml
[project]
name = "workshop-airflow"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "apache-airflow (>=3.1.3,<4.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "streamlit (>=1.51.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
